requirements.txt
# TODO: Split this file in at least tow(local, production) 
django==4.1.3
pytest==7.2.0
pytest-django==4.5.2
model_bakery==1.9.0
djangorestframework==3.14.0
coverage===6.5.0
